move verifyquote out of internal package
