+++ quote/quote.go
@@ -1,4 +1,4 @@
-package internal
+package quote
 
 import (
 	"bytes"
@@ -8,6 +8,7 @@ import (
 	"crypto/subtle"
 	"fmt"
 
+	"github.com/google/go-tpm-tools/internal"
 	pb "github.com/google/go-tpm-tools/proto/tpm"
 	"github.com/google/go-tpm/legacy/tpm2"
 )
@@ -16,7 +17,7 @@ import (
 // their preferred order of use.
 var SignatureHashAlgs = []tpm2.Algorithm{tpm2.AlgSHA512, tpm2.AlgSHA384, tpm2.AlgSHA256}
 
-// VerifyQuote performs the following checks to validate a Quote:
+// Verify performs the following checks to validate a Quote:
 //   - the provided signature is generated by the trusted AK public key
 //   - the signature signs the provided quote data
 //   - the quote data starts with TPM_GENERATED_VALUE
@@ -29,8 +30,8 @@ var SignatureHashAlgs = []tpm2.Algorithm{tpm2.AlgSHA512, tpm2.AlgSHA384, tpm2.Al
 // Note that the caller must have already established trust in the provided
 // public key before validating the Quote.
 //
-// VerifyQuote supports ECDSA and RSASSA signature verification.
-func VerifyQuote(q *pb.Quote, trustedPub crypto.PublicKey, extraData []byte) error {
+// Verify supports ECDSA and RSASSA signature verification.
+func Verify(q *pb.Quote, trustedPub crypto.PublicKey, extraData []byte) error {
 	sig, err := tpm2.DecodeSignature(bytes.NewBuffer(q.GetRawSig()))
 	if err != nil {
 		return fmt.Errorf("signature decoding failed: %v", err)
@@ -124,10 +125,10 @@ func verifyRSASSAQuoteSignature(rsaPub *rsa.PublicKey, hash crypto.Hash, quoted 
 }
 
 func validatePCRDigest(quoteInfo *tpm2.QuoteInfo, pcrs *pb.PCRs, hash crypto.Hash) error {
-	if !SamePCRSelection(pcrs, quoteInfo.PCRSelection) {
+	if !internal.SamePCRSelection(pcrs, quoteInfo.PCRSelection) {
 		return fmt.Errorf("given PCRs and Quote do not have the same PCR selection")
 	}
-	pcrDigest := PCRDigest(pcrs, hash)
+	pcrDigest := internal.PCRDigest(pcrs, hash)
 	if subtle.ConstantTimeCompare(quoteInfo.PCRDigest, pcrDigest) == 0 {
 		return fmt.Errorf("given PCRs digest not matching")
 	}
